feat: move annotations to new attribute namespace
